Make netrc deployment safer while not depending on deploy specifics
Both nix-daemon and hercules-ci users need access to a valid netrc.
Other users should not be able to access the netrc secrets. I can't
guarantee that nix-daemon doesn't expose it though.

Because both client and daemon need access to the netrc file, so
in order to avoid some confusion, the pull caches are only added
to the trustedSubstituters and thus need to be enabled specifically
by clients. Otherwise clients without netrc access generate warnings
and possibly nix-daemon state is affected by those. I have yet to
reconfirm the latter.

This also introduces a mechanism to add / override Nix options in
the agent, which may come in handy in the future.

The cachix-use-export.nix module can be improved for general
consumption, but is tailored to hercules-ci-agent for now.