agent: Dhall -> JSON, TOML
Resolves https://github.com/hercules-ci/hercules-ci-agent/issues/73

Dhall is a nice source language for configuration, but not a well-supported
target language. The agent is not using the powerful Dhall features like
functions and probably doesn't need to.

By providing JSON as a configuration language, we simplify most automation,
including Nix modules. Dhall can still be used, via dhall-to-json.

We also provide TOML support, which is conceptually close to JSON, supports
comments and isn't as fragile as YAML.