Split the API package
Duplications:
 - AgentInfo: this should probably diverge anyway, because the agent
      representation can't have the derived info like lastSeen, or
      anything about its load for example.
 - DerivationStatus: duplicated between frontend and agent-ifd code.
      Ifd doesn't need to know about the exact status, just about
      wait/succeed/fail, so that should be simplified anyway.
 - Message: duplicated between frontend and agent evaluation. I can't
      think of a reason why these should diverge, but it'd be the
      only real domain type in the core package, which doesn't seem
      right.

Id and Name phantoms are now poly-kinded to allow Symbols. Most objects
don't have a canonical type, so we should not pretend that they do.
By using Symbols instead, we avoid this and we'll need fewer imports,
which is good for the brain and probably for compiler performance too.