module: Make restarter more robust
These changes work around a systemd behavior where the unit is not
updated when it is trying to restart.
Restarts of the agent are usually caused by secrets file updates,
triggered by NixOps via file notifications and the restarter script.
By checking the configuration before attempting to restart, we
avoid this behavior.