Require that effects are below outputs.effects (for new format)
By requiring that effects are listed within an explicit attribute,
other expressions that are referenced outside in other places do
not get a chance to define an effect that will be run.

This reduces the "blast radius" of a compromised expression,
in cases where the compromised package (or other value) is only
used in a context that does not expose the compromised expression
or package to secrets. It can not gain access to secrets unless
it has been placed below effects.

Expressions that are only used as dependencies of derivations,
effects or shells already could not pretend to be effects.
An effect must be in an attribute that is exposed by `ci.nix`,
`flake.nix`, etc. Ie, a derivation that was not listed as an
"Attribute" in the dashboard would not have been able to run
as an effect regardless.