Revert "hercules-ci-cli: Update changelog"
The package does not pin its agent dependency, so it will use
the latest agent from hackage. This means that a fix in agent
will also be available in earlier versions and that the new
CLI release would be equal to the previous one.

That doesn't make sense, so no hci release.

This reverts commit 7cca85763e0849ed5f111ac0998288a42aa0f879.