Take charge of signal handling
It is now centralized in the `init` functions. Call either
- store only: Hercules.CNix.init, or
- store + expr: Hercules.CNix.Expr.init

This reduces the dependency in nix-main and improves signal handling
on macOS, as nix-main's initNix() would install signal handlers that
override the GHC runtime's signal handlers. That was worked around
by calling `installDefaultSigINTHandler`, but this commit solves it
properly as part of the `init` functions.

`installDefaultSigINTHandler` is therefore redundant, and may be
hidden in a later version of the bindings.

Finally this commit removes the dependency on `handleExceptions`, so
that the store bindings do not require nix-main anymore.